Update FileUpload.php (#114)
Fixed Type casting error

+++ src/FileUpload.php
@@ -34,11 +34,11 @@ class FileUpload extends \Contao\FileUpload
         $this->setName($name);
 
         $this->extensions = StringUtil::trimsplit(',', strtolower(Config::get('uploadTypes')));
-        $this->maxFileSize = Config::get('maxFileSize') ?? 0;
-        $this->imageWidth = Config::get('imageWidth') ?? 0;
-        $this->imageHeight = Config::get('imageHeight') ?? 0;
-        $this->gdMaxImgWidth = Config::get('gdMaxImgWidth') ?? 0;
-        $this->gdMaxImgHeight = Config::get('gdMaxImgHeight') ?? 0;
+        $this->maxFileSize = (int) Config::get('maxFileSize') ?? 0;
+        $this->imageWidth = (int) Config::get('imageWidth') ?? 0;
+        $this->imageHeight = (int) Config::get('imageHeight') ?? 0;
+        $this->gdMaxImgWidth = (int) Config::get('gdMaxImgWidth') ?? 0;
+        $this->gdMaxImgHeight = (int) Config::get('gdMaxImgHeight') ?? 0;
     }
 
     public function getName()
